# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: ChaloTogether API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /communities/{id}:
    get:
      operationId: getCommunity
      tags: [communities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Community detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /communities/{id}/members:
    get:
      operationId: getCommunityMembers
      tags: [communities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityMember"
  /communities/{id}/join:
    post:
      operationId: joinCommunity
      tags: [communities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Not verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /communities/{id}/leave:
    post:
      operationId: leaveCommunity
      tags: [communities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Events ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        demoToken:
          type: ["string", "null"]
        demoOtp:
          type: ["string", "null"]
...
    CommunityDetail:
      type: object
      required: [id, name, collegeId, collegeName, memberCount, isJoined, activeMembers, upcomingEvents, communityRides, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        collegeId:
          type: integer
        collegeName:
          type: string
        description:
          type: ["string", "null"]
        memberCount:
          type: integer
        activeMembers:
          type: integer
        upcomingEvents:
          type: integer
        communityRides:
          type: integer
        isJoined:
          type: boolean
        createdAt:
          type: string
    CommunityMember:
      type: object
      required: [id, fullName, college, department, year, isVerified, reliabilityScore, joinedAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        college:
          type: string
        department:
          type: string
        year:
          type: string
        profilePhotoUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        reliabilityScore:
          type: integer
        totalRides:
          type: integer
        joinedAt:
          type: string
...